Fix double-update in RectilinearOutputRateEncoder
Fix silent bug that happened when RectilinearOutputRateEncoder.update_rates was called twice in a row with the same update time. this happens commonly when an input sequence finishes (with a final rate update) to be immediately followed by andother action requiring a rate update. Now this situation is detected and the log index is decremented before update so as to overwrite the last record.